requirements.txt
asgiref==3.8.1
Django==4.2.16
django-unfold==0.39.0
sqlparse==0.5.1
typing_extensions==4.12.2
